compeletion of delete departs (set DepDeleteException)

+++ Intelligent learning system management/src/main/java/com/learn/service/impl/DeptServiceImpl.java
@@ -1,6 +1,8 @@
 package com.learn.service.impl;
 
+import com.learn.exception.DepDeleteException;
 import com.learn.mapper.DeptMapper;
+import com.learn.mapper.EmpMapper;
 import com.learn.pojo.Dept;
 import com.learn.service.DeptService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,6 +16,8 @@ public class DeptServiceImpl implements DeptService {
 
     @Autowired
     private DeptMapper deptMapper;
+    @Autowired
+    private EmpMapper empMapper;
 
     @Override
     public List<Dept> findAll() {
@@ -22,6 +26,11 @@ public class DeptServiceImpl implements DeptService {
 
     @Override
     public void deleteById(Integer id) {
+
+        int empCount = empMapper.countById(id);
+        if (empCount > 0) {
+            throw new DepDeleteException();
+        }
         deptMapper.deleteById(id);
     }
